2_ticket_triage/requirements.txt
faiss-cpu
numpy
sentence-transformers
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
streamlit
